shankarisr-3122225001125-connector/requirements.txt
# Minimal ETL Dependencies - Python 3.12 Compatible
requests>=2.25.0
pymongo>=4.0.0
python-dotenv>=0.19.0

# Optional: Only install if needed
pandas>=1.3.0
numpy>=1.21.0